// src/api-utils/generated-client/api.ts
export interface GetAccountDto {
    /**
     * 
     * @type {string}
     * @memberof GetAccountDto
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof GetAccountDto
     */
    'legal_name'?: string;
    /**
     * 
     * @type {string}
     * @memberof GetAccountDto
     */
    'federal_tax_id_no'?: string;
    /**
     * 
     * @type {string}
     * @memberof GetAccountDto
     */
    'legal_address'?: string;
    /**
     * 
     * @type {string}
     * @memberof GetAccountDto
     */
    'legal_address_city'?: string;
    /**
     * 
     * @type {string}
     * @memberof GetAccountDto
     */
    'legal_address_state'?: string;
    /**
     * 
     * @type {string}
     * @memberof GetAccountDto
     */
    'legal_address_zip'?: string;
    /**
     * 
     * @type {string}
     * @memberof GetAccountDto
     */
    'industry'?: string;
    /**
     * 
     * @type {string}
     * @memberof GetAccountDto
     */
    'website'?: string;
    /**
     * 
     * @type {number}
     * @memberof GetAccountDto
     */
    'annualrevenue'?: number;
    /**
     * 
     * @type {number}
     * @memberof GetAccountDto
     */
    'credit_amount_available'?: number;
    /**
     * 
     * @type {number}
     * @memberof GetAccountDto
     */
    'approved_credit_amount'?: number;
    /**
     * 
     * @type {string}
     * @memberof GetAccountDto
     */
    'document_signature_hash'?: string;
    /**
     * 
     * @type {string}
     * @memberof GetAccountDto
     */
    'deposit_bank_uuid'?: string;
    /**
     * 
     * @type {string}
     * @memberof GetAccountDto
     */
    'withdrawal_bank_uuid'?: string;
    /**
     * 
     * @type {number}
     * @memberof GetAccountDto
     */
    'total_outstanding_amount'?: number;
    /**
     * 
     * @type {string}
     * @memberof GetAccountDto
     */
    'payro_finance_status'?: GetAccountDtoPayroFinanceStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof GetAccountDto
     */
    'funding_status'?: GetAccountDtoFundingStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof GetAccountDto
     */
    'uuid'?: string;
    /**
     * 
     * @type {string}
     * @memberof GetAccountDto
     */
    'payroll_company'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof GetAccountDto
     */
    'contract_signed'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof GetAccountDto
     */
    'block_portal_access'?: boolean;
}

export const GetAccountDtoPayroFinanceStatusEnum = {
    New: 'New',
    Approved: 'Approved',
    OnHold: 'On Hold',
    Declined: 'Declined'
} as const;

export type GetAccountDtoPayroFinanceStatusEnum = typeof GetAccountDtoPayroFinanceStatusEnum[keyof typeof GetAccountDtoPayroFinanceStatusEnum];
export const GetAccountDtoFundingStatusEnum = {
    Yes: 'Yes',
    No: 'No'
} as const;

export type GetAccountDtoFundingStatusEnum = typeof GetAccountDtoFundingStatusEnum[keyof typeof GetAccountDtoFundingStatusEnum];
// ...
export interface GetRepaymentDto {
    /**
     * 
     * @type {number}
     * @memberof GetRepaymentDto
     */
    'bank_account_number': number;
    /**
     * 
     * @type {string}
     * @memberof GetRepaymentDto
     */
    'withdrawl_bank_name': string;
    /**
     * 
     * @type {number}
     * @memberof GetRepaymentDto
     */
    'expected_total_payback': number;
    /**
     * 
     * @type {object}
     * @memberof GetRepaymentDto
     */
    'full_remaining_balance_incl_interest': object;
    /**
     * 
     * @type {string}
     * @memberof GetRepaymentDto
     */
    'funded_date': string;
    /**
     * 
     * @type {number}
     * @memberof GetRepaymentDto
     */
    'funding_amount': number;
    /**
     * 
     * @type {string}
     * @memberof GetRepaymentDto
     */
    'next_hit_date': string;
    /**
     * 
     * @type {number}
     * @memberof GetRepaymentDto
     */
    'next_interest_amount': number;
    /**
     * 
     * @type {string}
     * @memberof GetRepaymentDto
     */
    'next_interest_date': string;
    /**
     * 
     * @type {string}
     * @memberof GetRepaymentDto
     */
    'payoff': string;
    /**
     * 
     * @type {string}
     * @memberof GetRepaymentDto
     */
    'payroll_record': string;
    /**
     * 
     * @type {object}
     * @memberof GetRepaymentDto
     */
    'payroll_record__r': object;
    /**
     * 
     * @type {string}
     * @memberof GetRepaymentDto
     */
    'status': string;
    /**
     * 
     * @type {string}
     * @memberof GetRepaymentDto
     */
    'salesforce_id': string;
    /**
     * 
     * @type {number}
     * @memberof GetRepaymentDto
     */
    'total_amount_settled': number;
    /**
     * 
     * @type {number}
     * @memberof GetRepaymentDto
     */
    'total_outstanding_amount': number;
    /**
     * 
     * @type {number}
     * @memberof GetRepaymentDto
     */
    'total_outstanding_interest': number;
    /**
     * 
     * @type {number}
     * @memberof GetRepaymentDto
     */
    'total_outstanding_principal': number;
    /**
     * 
     * @type {string}
     * @memberof GetRepaymentDto
     */
    'number_of_pay_periods': string;
    /**
     * 
     * @type {number}
     * @memberof GetRepaymentDto
     */
    'total_credits_settled': number;
    /**
     * 
     * @type {string}
     * @memberof GetRepaymentDto
     */
    'merchant_payroll_company': string;
    /**
     * 
     * @type {string}
     * @memberof GetRepaymentDto
     */
    'uuid': string;
}
/**
 * 
 * @export
 * @interface LoanPayment
 */
export interface LoanPayment {
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'repayment': string;
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'repayment_r.uuid__c': string;
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'repayment_account_name': string;
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'bank_name_from_repayment': string;
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'bank_account_from_repayment': string;
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'repayment_bank_account_number': string;
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'repayment_bank_name': string;
    /**
     * 
     * @type {number}
     * @memberof LoanPayment
     */
    'repayment_funding_amount': number;
    /**
     * 
     * @type {number}
     * @memberof LoanPayment
     */
    'running_balance': number;
    /**
     * 
     * @type {number}
     * @memberof LoanPayment
     */
    'running_fee_balance': number;
    /**
     * 
     * @type {number}
     * @memberof LoanPayment
     */
    'running_principal': number;
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'settlement_date': string;
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'effective_date': string;
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'status': string;
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'status_amount': string;
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'sub_type': string;
    /**
     * 
     * @type {number}
     * @memberof LoanPayment
     */
    'transaction_amount': number;
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'transaction_type': string;
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'uuid': string;
}
// ...
export interface NextPayroll {
    /**
     * 
     * @type {string}
     * @memberof NextPayroll
     */
    'payroll_date': string;
    /**
     * 
     * @type {number}
     * @memberof NextPayroll
     */
    'payroll_amount': number;
    /**
     * 
     * @type {string}
     * @memberof NextPayroll
     */
    'number_of_days_untill_payroll': string;
}